Format + clean code
